fix backup loop and rename backup.py to backup_helper.py
