+++ app/database/__init__.py
@@ -1,5 +1,5 @@
-from . import backup, db_helper
+from . import backup_helper, db_helper
 from .common import cache
 from .model import db
 
-__all__ = [db, backup, db_helper, cache]
+__all__ = [db, backup_helper, db_helper, cache]
